Add files via upload
Fine-tuning DistilBERT/BERT on FINRA dataset.
This is the first part of the process which consist of removing duplicate entries and text with word count at most 30 are removed. The training, validation and test data consist of the 'Topic' column and the preprocessed 'Text' column. The split data is saved as data_clean.