Add files via upload
Fine-tuning BERT on FINRA dataset. This is the third part of the process which consist of predicting and evaluating the BERT model performance on the test data. The test data consist of 18 entries. The model was trained on 50 epochs and the learning rate was set at 1e-5 and batch size 6. The best model was saved at epoch 31. The total time taken to create the data frame with top 3 rank of predictions in the test data is 0:00:43 (h:mm:ss). The macro-averaged f1-score on the test data is 1.00.